Refactor internal tests
